Refactor/getInfo: route.js - auth import 출처 표시 오류 코드 수정
